+++ changelog.md
@@ -7,6 +7,8 @@
 - Update `Invoice` type to actual API version
 - Update `createInvoice` method to actual API version, change related options types
 - Add `deleteInvoice` and `deleteCheck` methods
+- Change return type of `getExchangeRate` method and field `rate` of `ExchangeRate` type which returned by `getExchangeRates` method to string
+- Remove calculating and returning exchange rate for reverse pair of `source` and `target` by `getExchangeRate` method
 - Change `Balances` type to store available and on hold balances in new `Balance` type
 - Create `BalancesType` type to store available or on hold balances
 - Remove `isReturnInNanos` and `isForce` parameters from `getBalances` and `getBalance` methods